requirements.txt
PyYAML==6.0.1
python-hcl2==4.3.2
networkx==3.2.1
pytest==7.4.3
